search analytics tweaks
