+++ src/lib/utils.ts
@@ -221,6 +221,19 @@ export function downloadEvent(
 	analyticsEvent(eventData);
 }
 
+export function areaSelectEvent(area, interactionType, interactionValue) {
+	const isPostcode = area.type === 'postcode';
+	const eventData = {
+		event: 'interaction',
+		interactionType,
+		interactionValue,
+		areaCode: area.areacd,
+		areaName: area.areanm || area.areacd,
+		areaType: isPostcode ? 'postcode' : geoLevelsAllLookup?.[area.areacd.slice(0, 3)]?.label
+	};
+	analyticsEvent(eventData);
+}
+
 export function filterExtremeAreas(data: jsonDataCols, extremeAreas: string[] | null) {
 	if (!extremeAreas || !data.areacd) {
 		return data;
